add screen to allow retrying db connection

+++ 07-Database-JDBC/src/main/java/com/leevilaune/currency/dao/CurrencyDao.java
@@ -18,10 +18,16 @@ public class CurrencyDao {
 
     public Currency find(int id) {
         EntityManager em = MariaDbJpaConnection.getInstance();
+        if(em == null){
+            return null;
+        }
         return em.find(Currency.class, id);
     }
     public Currency find(String code){
         EntityManager em = MariaDbJpaConnection.getInstance();
+        if(em == null){
+            return null;
+        }
         return em.createQuery("select c from Currency c where c.code = :code", Currency.class)
                 .setParameter("code", code)
                 .getSingleResult();
@@ -29,12 +35,18 @@ public class CurrencyDao {
 
     public List<Currency> findAll() {
         EntityManager em = MariaDbJpaConnection.getInstance();
+        if(em == null){
+            return null;
+        }
         List<Currency> emps = em.createQuery("select c from Currency c").getResultList();
         return emps;
     }
 
     public void update(Currency curr) {
         EntityManager em = MariaDbJpaConnection.getInstance();
+        if(em == null){
+            return;
+        }
         em.getTransaction().begin();
         em.merge(curr);
         em.getTransaction().commit();
@@ -42,6 +54,9 @@ public class CurrencyDao {
 
     public void delete(Currency curr) {
         EntityManager em = MariaDbJpaConnection.getInstance();
+        if(em == null){
+            return;
+        }
         em.getTransaction().begin();
         em.remove(curr);
         em.getTransaction().commit();
